pyproject.toml
[tool.poetry]
name = "motion-Pattern-analysis-orthopedic"
version = "0.1.0"
description = "A tool for analyzing knee angles from video footage"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "knee_analysis"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
opencv-python = "^4.8.1.78"
mediapipe = "^0.10.8"
numpy = "^1.26.3"
matplotlib = "^3.8.2"
scipy = "^1.15.3"
pygame = "^2.5.2"
PyOpenGL = "^3.1.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
knee-analyzer = "knee_analysis.knee_angle_analyzer:main"
